demo_prefect_flow_consume_t212_api/requirements.txt
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
apprise==1.9.5
asgi-lifespan==2.1.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.4.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
coolname==2.2.0
cryptography==46.0.2
dateparser==1.2.2
docker==7.1.0
exceptiongroup==1.3.0
fastapi==0.118.2
fsspec==2025.9.0
graphviz==0.21
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
oauthlib==3.3.1
opentelemetry-api==1.37.0
orjson==3.11.3
packaging==25.0
pathspec==0.12.1
pendulum==3.1.0
prefect==3.4.23
prefect-docker==0.6.6
prometheus_client==0.23.1
pycparser==2.23
pydantic==2.12.0
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
python-socks==2.7.2
pytz==2025.2
PyYAML==6.0.3
readchar==4.2.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.2.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlitedict==2.1.0
sqlmodel==0.0.27
starlette==0.48.0
text-unidecode==1.3
toml==0.10.2
tomli==2.3.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uv==0.9.1
uvicorn==0.37.0
websockets==15.0.1
zipp==3.23.0
